refactor(search-input): rename variables and remove unnecessary ESLint rule disable
