+++ src/components/search-input/functions.ts
@@ -1,6 +1,6 @@
 import { Hit } from 'react-instantsearch-core'
 
-export const breadcrumbs = (hit: Hit) => {
+export const getBreadcrumbs = (hit: Hit) => {
   const urlSplit = hit.url.split('/').slice(5)
   const endpointSplit = urlSplit.map((e) => e.split('#'))
   return endpointSplit.flat()
